Safari-611.1.10/LayoutTests/js/invalid-utf8-in-syntax-error-expected.txt

Ensures that we correctly propagate the error message for lexer errors containing invalid utf8 code sequences

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ({f("�")}) threw exception SyntaxError: Unexpected string literal "�". Expected a parameter pattern or a ')' in parameter list..
PASS successfullyParsed is true

TEST COMPLETE
